mv TransformMetaTable to step.batch_transform
